+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
+export { default as useCartAction } from './useCartAction';
 export { default as useFetch } from './useFetch';
 export { default as useStackProducts } from './useStackProducts';
 export { default as useFilteredProducts } from './useFilteredProducts';
